Orginise test folder
