+++ NanoEngine/Testing/Assets/TestAsset.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace NanoEngine
+namespace NanoEngine.Testing.Assets
 {
     class TestAsset : Entity
     {
