+++ NanoEngine/Testing/Tiles/GrassTile.cs
@@ -7,7 +7,7 @@ using Microsoft.Xna.Framework.Graphics;
 using NanoEngine.Core.Managers;
 using NanoEngine.ObjectTypes.Assets;
 
-namespace NanoEngine.Testing
+namespace NanoEngine.Testing.Tiles
 {
     class GrassTile : Tile
     {
